third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/softplus.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] softplus float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0.0078125 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 9633
[FAIL] softplus float16 0D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 14943
[FAIL] softplus float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 17825
[FAIL] softplus float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 17825
[FAIL] softplus float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 17825
[FAIL] softplus float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test softplus float16 actual 0 should be close enough to expected 5.62890625 by ULP distance: expected a number less than or equal to 18 but got 17825
Harness: the test ran to completion.
